+++ src/features/Trends/containers/index.tsx
@@ -1,0 +1,4 @@
+import { DesktopTrends } from "./DesktopTrends";
+import { MobileTrends } from "./MobileTrends";
+
+export { DesktopTrends, MobileTrends };