+++ src/features/Trends/index.tsx
@@ -1,0 +1,3 @@
+import { DesktopTrends, MobileTrends } from "./containers";
+
+export { DesktopTrends, MobileTrends };